Speed up the evaluation of the isochrone and spherical inverses
Profiling the interpolated spherical inverse showed that its evaluation
was dominated by the solve of the Kepler-like equation in
actionAngleIsochroneInverse, which looped over the phase-space points
in Python and called scipy's newton (with its isclose-based convergence
check) for each one.

- Solve that equation for all points at once with the safeguarded
  Newton-Raphson/bisection solver introduced for the spherical inverse,
  which is moved to actionAngleIsochroneInverse (the forward isochrone
  transformation obtains the eccentric anomaly directly and never has
  to solve this equation, while both inverse transformations do). This
  is also more accurate (converged to 1e-14 rather than scipy's default
  1.5e-8) and does not need the brentq fallback.
- Reuse a single isochrone-inverse instance in the interpolated
  evaluation instead of constructing one, together with an
  IsochronePotential that parses the configuration file, per call.
- Skip the interpolation of the mapping coefficients when evaluating
  with the point transformation only, where they are not used.

For a 31x31 grid of tori evaluated at 1000 angles along a torus, the
cost per phase-space point drops from 143 to 7.4 microseconds
(pt_only) and from 153 to 17.4 microseconds (with the
generating-function mapping); the fixed cost per call drops from about
1.1 to 0.65 milliseconds.
